Improve docs and refactor methods
