+++ modules/core/src/main/java/org/vstu/compprehension/models/businesslogic/domains/helpers/meaningtree/MeaningTreeDefaultExpressionConfig.java
@@ -2,6 +2,9 @@ package org.vstu.compprehension.models.businesslogic.domains.helpers.meaningtree
 
 import java.util.HashMap;
 
+/**
+ * Base config for Meaning Tree required for correct Expression domain functioning
+ */
 public class MeaningTreeDefaultExpressionConfig extends HashMap<String, String> {
     public MeaningTreeDefaultExpressionConfig() {
         super();
